Added special cases for 0 or 1 legal move at root. Changed version number to 0.2 beta.

+++ search.h
@@ -36,7 +36,7 @@ const int DEPTH = 2;
 // const int NODES = 3;
 
 // Time constants
-const int ONE_SECOND = 1000;
+const uint64_t ONE_SECOND = 1000;
 const uint64_t MAX_TIME = (1ULL << 63) - 1;
 
 // Search parameters
